feature 1897 added: the core "Contact Webmaster" link in footer is replaced by a link to the ContactForm page. It works with a smarty prefilter, so it should work with all themes.
The configurable template variable name was removed. It is hard coded as
$ContactFormLink. I've seen no example using another variable name. It was
hard to understand for users in the administration panel and not very useful.
The corresponding language keys were removed from all languages.

bug 1916 fixed: apply the same rule for link display in footer as in menu. If
the administrator has allowed the ContactForm page for guests, let's display
it in the menu + footer.


git-svn-id: http://piwigo.org/svn/extensions/ContactForm@7197 68402e56-0260-453c-a942-63ccdbb3a9ee